requirements.txt
fastapi==0.104.1
google-cloud-secret-manager==2.22.0
python-dotenv==1.0.0
google-auth==2.23.4